separate mask creation from section transport computation
